+++ client/views/pages/browse_collection/browse_collection.html
@@ -10,6 +10,12 @@
                                 <small></small>
                             </h5>
                             <div class="ibox-tools">
+                                <button id="btnExportQueryResult" type="button" class="btn btn-sm btn-primary"
+                                        data-target="#exportQueryResultModal"
+                                        data-toggle="modal"
+                                        style="display: none">Export Result
+                                </button>
+
                                 <a id="btnSwitchView" href="#"
                                    title="Switch Result View">
                                     <i class="fa fa-exchange"></i>
@@ -57,7 +63,9 @@
 
                                 <div class="form-group">
                                     <div class="col-lg-12">
-                                        <button id="btnExecuteQuery" class="btn btn-block btn-outline btn-primary ladda-button" type="button" data-style="contract"><strong>Execute</strong></button>
+                                        <button id="btnExecuteQuery"
+                                                class="btn btn-block btn-outline btn-primary ladda-button" type="button"
+                                                data-style="contract"><strong>Execute</strong></button>
                                     </div>
                                 </div>
                                 <ul id="resultTabs" class="nav nav-tabs">
@@ -91,5 +99,30 @@
         </div>
 
         {{> queryHistories}}
+
+        <!-- MODAL Export Query Result-->
+        <div class="modal inmodal" id="exportQueryResultModal" tabindex="-1" role="dialog" aria-hidden="true"
+             style="display: none;">
+            <div class="modal-dialog modal-sm">
+                <div class="modal-content animated bounceInRight">
+                    <div class="modal-header">
+                        <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">×</span><span
+                                class="sr-only">Close</span></button>
+                        <h4 class="modal-title">Choose Format</h4>
+                    </div>
+                    <div class="modal-footer">
+                        <div class="pull-left">
+                            <button id="btnExportAsCSV" type="button" class="btn btn-primary" data-dismiss="modal">CSV
+                            </button>
+                        </div>
+                        <div class="pull-right">
+                            <button id="btnExportAsJSON" type="button" class="btn btn-primary" data-dismiss="modal">
+                                JSON
+                            </button>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
     {{/if}}
 </template>
